Big refactor.
- Class Integer was removed.
- Added macros to guide addition and multiplication under cuFFT/NTT operation mode.

Ok for addition and multiplication under cuFFT mode.

+++ src/distribution.h
@@ -17,7 +17,7 @@ enum kind_t
 };
 
 #define MAX_DEGREE 16384
-#define SEED (unsigned long)(7226776555987513888)
+#define SEED (unsigned long long)(7226776555987513888)
 
 class Distribution{
   private:
